src/requirements/base.txt
# Core
Django>=1.8,<1.9
Markdown==2.6.2
django-filter==0.11.0
virtualenv==13.1.0
psycopg2==2.4.5
django-cors-headers==1.1.0
django-rest-swagger==0.3.4
djangorestframework==3.2.3
djangorestframework-jwt==1.7.2
drf-nested-routers==0.10.0
gunicorn==19.3.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
djrill==2.0
Pillow==2.3.0

# Packages
ipython
django-extensions
Werkzeug
